LPD-88085 move to util package
