+++ modules/apps/frontend-token/frontend-token-definition-api/src/main/java/com/liferay/frontend/token/definition/util/FrontendTokenDefinitionUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.frontend.token.definition;
+package com.liferay.frontend.token.definition.util;
 
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONException;
