+++ modules/apps/frontend-token/frontend-token-definition-api/src/main/resources/com/liferay/frontend/token/definition/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
